+++ src/main/scala/com/testerhome/appcrawler/CrawlerConf.scala
@@ -95,11 +95,17 @@ class CrawlerConf {
   /** 引导规则. name, value, times三个元素组成 */
   var triggerActions = ListBuffer[scala.collection.mutable.Map[String, Any]]()
   //todo: 用watch代替triggerActions
-  var watch = ListBuffer[Step]()
-  var startupActions = ListBuffer[String]()
+  var autoCrawl: Boolean=true
   var asserts = ListBuffer[Map[String, Any]]()
-  var testcase=TestCase()
-  startupActions ++= List("println(\"startupAction call use scala code\")")
+  var testcase=TestCase(
+    name="TesterHome AppCrawler",
+    steps = List(
+      Step(given = null, when = null, xpath="//*", action = "driver.swipe(0.9, 0,5, 0.1, 0.5)", then=null),
+      Step(given = null, when = null, xpath="//*", action = "driver.swipe(0.9, 0,5, 0.1, 0.5)", then=null),
+      Step(given = null, when = null, xpath="//*", action = "driver.swipe(0.9, 0,5, 0.1, 0.5)", then=null),
+      Step(given = null, when = null, xpath="//*", action = "driver.swipe(0.9, 0,5, 0.1, 0.5)", then=null)
+    )
+  )
 
   var beforeElementAction = ListBuffer[Map[String, String]]()
   var afterElementAction = ListBuffer[String]()