+++ src/main/scala/com/testerhome/appcrawler/driver/AppiumClient.scala
@@ -201,7 +201,6 @@ class AppiumClient extends CommonLog with WebBrowser with WebDriver{
     if (resultDir.nonEmpty) {
       crawler.conf.resultDir = resultDir
     }
-    crawler.conf.startupActions.clear()
     crawler.log.setLevel(Level.TRACE)
     crawler.conf.maxDepth = maxDepth
     crawler.start(this)